+++ previous-versions/assess/codatassess.go
@@ -134,8 +134,8 @@ func New(opts ...SDKOption) *CodatAssess {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "1.0",
-			SDKVersion:        "0.26.0",
-			GenVersion:        "2.129.1",
+			SDKVersion:        "0.26.1",
+			GenVersion:        "2.131.1",
 		},
 	}
 	for _, opt := range opts {